feat(simple-iam-server-starter): 发布1.1.0，详见CHANGELOG.1.1.0.md

+++ sdk/auth/iam/server/simple-iam-server-starter/src/main/java/io/github/surezzzzzz/sdk/auth/iam/server/dto/message/response/MessageSendResponse.java
@@ -14,5 +14,8 @@ public class MessageSendResponse {
 
     private String sendBatchId;
 
-    private int recipientCount;
+    /**
+     * 实际收件人数，面向全量发送结果使用 long，避免人数增长后的计数溢出。
+     */
+    private long recipientCount;
 }
